Refactor to persist multiple reactflow instances
