+++ public/pages/workflow_detail/workspace/workspace-styles.scss
@@ -2,5 +2,4 @@
   // this ratio will allow the workspace to render on a standard
   // laptop (3024 x 1964) without introducing overflow/scrolling
   height: 60vh;
-  padding: 0;
 }
