+++ public/pages/workflow_detail/workspace_component/output_handle.tsx
@@ -3,12 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useState, useRef, useEffect, useContext } from 'react';
-import { Connection, Handle, Position } from 'reactflow';
+import React, { useState, useRef, useEffect } from 'react';
+import { Connection, Handle, Position, useReactFlow } from 'reactflow';
 import { EuiText } from '@elastic/eui';
 import { IComponent, IComponentOutput } from '../../../component_types';
 import { calculateHandlePosition, isValidConnection } from './utils';
-import { rfContext } from '../../../store';
 
 interface OutputHandleProps {
   data: IComponent;
@@ -17,7 +16,7 @@ interface OutputHandleProps {
 
 export function OutputHandle(props: OutputHandleProps) {
   const ref = useRef(null);
-  const { reactFlowInstance } = useContext(rfContext);
+  const reactFlowInstance = useReactFlow();
   const [position, setPosition] = useState<number>(0);
   const outputClasses = props.output.baseClasses.join('|');
