+++ public/render_app.tsx
@@ -9,7 +9,7 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { AppMountParameters, CoreStart } from '../../../src/core/public';
 import { FlowFrameworkDashboardsApp } from './app';
-import { store, ReactFlowContextProvider } from './store';
+import { store } from './store';
 
 export const renderApp = (
   coreStart: CoreStart,
@@ -17,13 +17,9 @@ export const renderApp = (
 ) => {
   ReactDOM.render(
     <Provider store={store}>
-      <ReactFlowContextProvider>
-        <Router basename={appBasePath + '#/'}>
-          <Route
-            render={(props) => <FlowFrameworkDashboardsApp {...props} />}
-          />
-        </Router>
-      </ReactFlowContextProvider>
+      <Router basename={appBasePath + '#/'}>
+        <Route render={(props) => <FlowFrameworkDashboardsApp {...props} />} />
+      </Router>
     </Provider>,
     element
   );
