--- public/store/context/index.ts
@@ -1,6 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './react_flow_context_provider';
